refactor code for maintainability
